1.统一loadingview共有属性；
2.隐藏不必要view；
3.修改闹钟loadingview秒针旋转中心位置；

+++ loadingview/src/main/java/com/fatdoge/loadingview/utils/DisplayUtils.java
@@ -10,44 +10,45 @@ public class DisplayUtils {
 
     /**
      * convert px to its equivalent dp
-     *
-     * 将px转换为与之相等的dp
+     * @param context
+     * @param pxValue
+     * @return
      */
-    public static int px2dp(Context context, float pxValue) {
+    public static float px2dp(Context context, float pxValue) {
         final float scale =  context.getResources().getDisplayMetrics().density;
-        return (int) (pxValue / scale + 0.5f);
+        return (pxValue / scale + 0.5f);
     }
-
 
     /**
      * convert dp to its equivalent px
-     *
-     * 将dp转换为与之相等的px
+     * @param context
+     * @param dipValue
+     * @return
      */
-    public static int dp2px(Context context, float dipValue) {
+    public static float dp2px(Context context, float dipValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
-        return (int) (dipValue * scale + 0.5f);
+        return (dipValue * scale + 0.5f);
     }
-
 
     /**
      * convert px to its equivalent sp
-     *
-     * 将px转换为sp
+     * @param context
+     * @param pxValue
+     * @return
      */
-    public static int px2sp(Context context, float pxValue) {
+    public static float px2sp(Context context, float pxValue) {
         final float fontScale = context.getResources().getDisplayMetrics().scaledDensity;
-        return (int) (pxValue / fontScale + 0.5f);
+        return (pxValue / fontScale + 0.5f);
     }
-
 
     /**
      * convert sp to its equivalent px
-     *
-     * 将sp转换为px
+     * @param context
+     * @param spValue
+     * @return
      */
-    public static int sp2px(Context context, float spValue) {
+    public static float sp2px(Context context, float spValue) {
         final float fontScale = context.getResources().getDisplayMetrics().scaledDensity;
-        return (int) (spValue * fontScale + 0.5f);
+        return (spValue * fontScale + 0.5f);
     }
 }
